feat(integrations): extract shared connector primitives
