+++ apps/web/package.json
@@ -15,6 +15,7 @@
     "@axe-core/playwright": "^4.11.2",
     "@dpf/db": "workspace:*",
     "@dpf/finance-templates": "workspace:*",
+    "@dpf/integration-shared": "workspace:*",
     "@dpf/storefront-templates": "workspace:*",
     "@dpf/validators": "workspace:*",
     "@floating-ui/react": "^0.27.19",
