+++ packages/integration-shared/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@dpf/integration-shared",
+  "version": "0.0.1",
+  "private": true,
+  "license": "Apache-2.0",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "scripts": {
+    "test": "vitest run",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.3",
+    "vitest": "^4.0.0"
+  }
+}
